backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
